Refactored the router in it's own class
